Moves settings from InspectorControl into the Sidebar
