+++ packages/edit-navigation/src/components/sidebar/auto-add-pages.js
@@ -1,8 +1,8 @@
 /**
  * WordPress dependencies
  */
-import { ToggleControl } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
+import { ToggleControl } from '@wordpress/components';
 
 /**
  * Internal dependencies
