+++ packages/edit-navigation/src/components/sidebar/menu-settings.js
@@ -1,0 +1,20 @@
+/**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+import { PanelBody } from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import { NameEditor } from '../name-editor';
+import AutoAddPages from './auto-add-pages';
+
+export default function MenuSettings( { menuId } ) {
+	return (
+		<PanelBody title={ __( 'Menu settings' ) }>
+			<NameEditor />
+			<AutoAddPages menuId={ menuId } />
+		</PanelBody>
+	);
+}
