+++ packages/edit-navigation/src/components/sidebar/style.scss
