+++ packages/edit-navigation/src/style.scss
@@ -12,7 +12,7 @@ $navigation-editor-spacing-top: $grid-unit-50 * 2;
 @import "./components/editor/style.scss";
 @import "./components/error-boundary/style.scss";
 @import "./components/header/style.scss";
-@import "./components/inspector-additions/style.scss";
+@import "./components/sidebar/style.scss";
 @import "./components/layout/style.scss";
 @import "./components/menu-switcher/style.scss";
 @import "./components/notices/style.scss";
